Fix: Refactor widget code and fix issues with dropdown [TMZ-811]
