+++ classes/renderable.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace HelloPlus\Classes;
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit; // Exit if accessed directly.
+}
+
+interface Renderable {
+
+	public function render(): void;
+}
