+++ modules/template-parts/assets/scss/hello-plus-header.scss
@@ -727,6 +727,10 @@
 					display: none;
 				}
 
+				& .ehp-header__contact-buttons {
+					justify-content: center;
+				}
+
 				.ehp-header__menu-cart {
 					display: none;
 				}
