+++ modules/template-parts/widgets/ehp-header.php
@@ -200,7 +200,6 @@ class Ehp_Header extends Ehp_Widget_Base {
 		$menus = $this->get_available_menus();
 
 		if ( ! empty( $menus ) ) {
-
 			$this->add_control(
 				'navigation_menu',
 				[
@@ -218,27 +217,6 @@ class Ehp_Header extends Ehp_Widget_Base {
 					),
 				]
 			);
-
-			foreach ( $this->get_empty_menus() as $menu_id => $menu_slug ) {
-
-				$this->add_control(
-					'empty_menu_alert_' . $menu_id,
-					[
-						'type' => Controls_Manager::ALERT,
-						'alert_type' => 'info',
-						'content' => sprintf(
-							/* translators: 1: Link opening tag, 2: Link closing tag. */
-							esc_html__( 'This menu has no items. Select another menu, or %1$sadd items%2$s. Then refresh this page. ', 'hello-plus' ),
-							sprintf( '<a href="%s" target="_blank">', self_admin_url( 'nav-menus.php?action=edit&menu=' . $menu_id ) ),
-							'</a>'
-						),
-						'separator' => 'after',
-						'condition' => [
-							'navigation_menu' => $menu_slug,
-						],
-					]
-				);
-			}
 
 			$this->add_control(
 				'navigation_icon_label',
@@ -888,6 +866,15 @@ class Ehp_Header extends Ehp_Widget_Base {
 		);
 
 		$this->add_control(
+			'style_responsive_menu_alert',
+			[
+				'type' => Controls_Manager::ALERT,
+				'alert_type' => 'info',
+				'content' => esc_html__( 'To preview, select a responsive viewport icon.', 'hello-plus' ),
+			]
+		);
+
+		$this->add_control(
 			'style_responsive_menu_align',
 			[
 				'label' => esc_html__( 'Text Align', 'hello-plus' ),
@@ -1242,9 +1229,6 @@ class Ehp_Header extends Ehp_Widget_Base {
 				],
 				'default' => 'navbar',
 				'separator' => 'before',
-				'condition' => [
-					'navigation_breakpoint' => [ 'mobile-portrait', 'tablet-portrait' ],
-				],
 			]
 		);
